move e2e tests to 01-e2eencryption
